Parsing LLMSetInput instead of LLMSetField
