+++ Stitch/Graph/Toolbar/View/CatalystNavigationBarHelperViews.swift
@@ -135,12 +135,12 @@ struct CatalystTopBarGraphButtons: View {
         // `HStack` doesn't matter? These are all placed in a `ToolbarItemGroup` ...
         HStack {
             
-            CatalystNavBarButton("rectangle.and.pencil.and.ellipsis") {
+            CatalystNavBarButton(LLM_OPEN_JSON_ENTRY_MODAL_SF_SYMBOL) {
                 dispatch(LLMActionsJSONEntryModalOpened())
             }
             .opacity(llmRecordingModeEnabled ? 1 : 0)
             
-            CatalystNavBarButton(llmRecordingModeActive ? "stop.fill" : "play.fill") {
+            CatalystNavBarButton(llmRecordingModeActive ? LLM_STOP_RECORDING_SF_SYMBOL : LLM_START_RECORDING_SF_SYMBOL) {
                 dispatch(LLMRecordingToggled())
             }
             .opacity(llmRecordingModeEnabled ? 1 : 0)
